new_project/requirements.txt
google-generativeai==0.8.6
google-ai-generativelanguage==0.6.15

chromadb==0.4.22
chroma-hnswlib==0.7.3
faiss-cpu==1.8.0.post1
posthog==3.5.0

fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.8.0
starlette==0.37.2

httpx==0.28.1
aiohttp==3.13.3
anyio==4.12.1

regex==2026.1.15
PyYAML==6.0.3
python-dotenv==1.2.1

orjson==3.10.6
tenacity==8.5.0
tqdm==4.66.4

python-docx==1.1.2
